Fix ambient lighting
Was broken because of a copy-paste fail in the way I set the ambient, diffuse and specular light colors. I happened to be setting the diffuse and specular colors to their defaults anyway, so the only difference was ambient light being zeroed out (default is (0,0,0,1) rather than (1,1,1,1) for the other components).

Have added screenshots of broken and correct behavior. Will move these screenshots to their own directory in the next commit.